+++ app/services/voice/VoiceVerificationResult.scala
@@ -1,10 +1,8 @@
 package services.voice
 
-import services.Utils.Enum
-
 trait VoiceVerificationResult {
   def success: Boolean
-  def score: Int
+  def score: Long
 }
